+++ common/src/main/java/net/anvian/sodiumextrainformation/util/RGB.java
@@ -1,0 +1,15 @@
+package net.anvian.sodiumextrainformation.util;
+
+public class RGB {
+    int r, g, b;
+
+    public RGB() {
+        this.r = 255;
+        this.g = 255;
+        this.b = 255;
+    }
+
+    public int rgbToDecimal() {
+        return (r << 16) + (g << 8) + b;
+    }
+}