+++ common/src/main/java/net/anvian/sodiumextrainformation/client/SodiumExtraInformationClientMod.java
@@ -3,7 +3,6 @@ package net.anvian.sodiumextrainformation.client;
 import net.anvian.anvianslib.config.TelemetryConfigManager;
 import net.anvian.anvianslib.util.LibUtil;
 import net.anvian.sodiumextrainformation.options.SodiumExtraInformationGameOptions;
-import net.anvian.sodiumextrainformation.platform.Services;
 import net.anvian.sodiumextrainformation.util.SessionManager;
 import net.caffeinemc.mods.sodium.client.services.PlatformRuntimeInformation;
 import org.slf4j.Logger;
@@ -36,18 +35,8 @@ public class SodiumExtraInformationClientMod {
     }
 
     public static void onInitialization() {
-        LibUtil.generateConfigPath(SodiumExtraInformationClientMod.MOD_ID, PlatformRuntimeInformation.getInstance().getConfigDirectory());
-
-        TelemetryConfigManager.initialize(PlatformRuntimeInformation.getInstance().getConfigDirectory().resolve(SodiumExtraInformationClientMod.MOD_ID).toFile());
-        if (TelemetryConfigManager.getConfig().enableTelemetry){
-            TelemetryConfigManager.sendTelemetryData(
-                    SodiumExtraInformationClientMod.MOD_ID,
-                    SodiumExtraInformationClientMod.MOD_VERSION,
-                    LibUtil.getMinecraftVersion(),
-                    Services.PLATFORM.getPlatformName(),
-                    !Services.PLATFORM.isDevelopmentEnvironment()
-            );
-        }
+        LibUtil.generateConfigPath(MOD_ID, PlatformRuntimeInformation.getInstance().getConfigDirectory());
+        TelemetryConfigManager.sendTelemetryData(MOD_ID, MOD_VERSION);
 
         CONFIG = SodiumExtraInformationGameOptions.load();
     }