+++ CHANGELOG.md
@@ -1,1 +1,1 @@
-- Updating dependencies to work in new versions
+- Add Anvian's Library integration