+++ CHANGELOG.md
@@ -1,3 +1,1 @@
-- Add Anvian's Library integration 
-- Config file was refactored
-- Every text can have a custom color
+- Update to 1.21.5